Fix format specifier warnings for ARM64 compatibility
- SoundManager: Use %u instead of %lu for UInt32
- MenuManager: Use int instead of size_t for %.*s precision
- TextEdit: Use %d with (int) casts for SInt32 values
- TextEdit: Use %u instead of %lu for UInt32

+++ src/MenuManager/MenuManagerCore.c
@@ -666,7 +666,7 @@ MenuHandle NewMenu(short menuID, ConstStr255Param menuTitle)
 {
     MenuHandle theMenu;
     MenuInfo* menuPtr;
-    size_t titleLen;
+    int titleLen;
 
     if (!gMenuMgrInitialized) {
         return NULL;
